+++ src/TgBotLib/Exceptions/NotImplementedException.php
@@ -1,0 +1,13 @@
+<?php
+
+    namespace TgBotLib\Exceptions;
+
+    use Throwable;
+
+    class NotImplementedException extends \Exception
+    {
+        public function __construct(string $message = "", int $code = 0, ?Throwable $previous = null)
+        {
+            parent::__construct($message, $code, $previous);
+        }
+    }